src/shared/security/aes.go
package security

import (
	"crypto/aes"
	"crypto/cipher"
	"crypto/rand"
	"encoding/base64"
	"errors"
	"fmt"
	"io"
)

// AES struct
type AES struct {
	key string
}

// NewAES Security AES's constructor
func NewAES(key string) *AES {
	return &AES{key}
}

// Encrypt string to base64 crypto using AES
func (a *AES) Encrypt(text string) (string, error) {
	// key := []byte(keyText)
	plaintext := []byte(text)

	keyByte := []byte(a.key)

	block, err := aes.NewCipher(keyByte)
	if err != nil {
		return "", err
	}

	// The IV needs to be unique, but not secure. Therefore it's common to
	// include it at the beginning of the ciphertext.
	ciphertext := make([]byte, aes.BlockSize+len(plaintext))
	iv := ciphertext[:aes.BlockSize]
	if _, err := io.ReadFull(rand.Reader, iv); err != nil {
		return "", err
	}

	stream := cipher.NewCFBEncrypter(block, iv)
	stream.XORKeyStream(ciphertext[aes.BlockSize:], plaintext)

	// convert to base64
	return base64.URLEncoding.EncodeToString(ciphertext), nil
}

// Decrypt from base64 to decrypted string
func (a *AES) Decrypt(encryptedText string) (string, error) {
	ciphertext, _ := base64.URLEncoding.DecodeString(encryptedText)

	keyByte := []byte(a.key)

	block, err := aes.NewCipher(keyByte)
	if err != nil {
		return "", err
	}

	// The IV needs to be unique, but not secure. Therefore it's common to
	// include it at the beginning of the ciphertext.
	if len(ciphertext) < aes.BlockSize {
		return "", errors.New("ciphertext too short")
	}
	iv := ciphertext[:aes.BlockSize]
	ciphertext = ciphertext[aes.BlockSize:]

	stream := cipher.NewCFBDecrypter(block, iv)

	// XORKeyStream can work in-place if the two arguments are the same.
	stream.XORKeyStream(ciphertext, ciphertext)

	return fmt.Sprintf("%s", ciphertext), nil
}
